chore: Move `createFileReloader` into it's own file
